requirements.txt
python-telegram-bot
opencv-python-headless
numpy
requests
easyocr
